adding USIC tests -- local only
